Refactor so that deletion of couchdb collection happens when synced.
